requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=1.4
asyncpg
python-dotenv
